Enable re-use of local arrays as seeds inside delayed arrays.
Any delayed array can now define a 'custom alabaster local array' as a
seed, which will instruct the loaders in this package to look for a
array in the same project and use that as the seed. This enables some
efficiencies when multiple delayed arrays share the same seed.

Note that the default stagers aren't capable of actually creating such
custom arrays. From the R side, we can't easily detect whether one array
is a duplicate of another that was encountered in a previous stageArray
call. If you want to take advantage of this, you have to stage manually.

Also a breaking change to createRawArraySeed so that it accepts project=
rather than a path= to the file. This should make it easier to use as
any caller would otherwise be calling acquireFile manually.