Tighted DelayedMask behavior to align with missing value policies.
The major change is that, if 'is.na()' is TRUE for a placeholder, we will mask
all values for which 'is.na()' is TRUE; this includes NaNs that were previously
not masked. This aligns with the takane policy of not trusting NaN payloads.

Another change is that, if we see an integer dataset with a NA and the
placeholder is not NA, we throw an error because the dataset's NA was likely an
attempt at -2^31, which we just can't represent as a regular integer in R.