Iterate according to chunks of the output HDF5 dataset.
This is the lesser of the two evils, as the input dataset might not need
consideration of chunks; whereas we know for sure that the output is HDF5-based
and thus would be expensive to breach chunk boundaries.

We also fix the block-writing tests so that we actually use small chunks
and small blocks, so as to encourage non-trivial iteration.