Use altSaveObject to save pristine seeds in a DelayedArray.
This ensures that we respond to any application overrides. However, it
requires some more care as altSaveObject may not be a generic (and even
if it is, it might just delegate to saveObject in its ANY method). So if
there is an override, we have to just try() it and see if it works...
and if it doesn't, we emit a warning if the error is that of dispatch.

That said, if there is no override, we continue to use selectMethod as
this is less nasty than attempting to parse the error message.