Deprecate WrapperArraySeed in favor of DelayedUnaryIsoOp.
Now ReloadedArraySeed directly inherits from DelayedUnaryIsoOp, which already
implements all of the trivial methods that forward the operation to the seed.

The only operation that isn't supported is the dedicated coercion methods from
to dgCMatrix. Turns out it wasn't that useful anyway, because there is no
longer a dedicated fast path from CSC_H5SparseMatrixSeeds to dgCMatrix.